refactor(packages/eg-walker): remove re-exports from non-index files
- Remove GraphEvent re-export from event-graph.ts
- Remove wildcard re-export from compat.ts
- Update all imports to reference original source (../types)
- Update graph/index.ts to match removed exports

All 300 tests passing, 0 TypeScript errors
