+++ packages/eg-walker/src/graph/event-graph.ts
@@ -7,9 +7,6 @@
 
 import type { GraphEvent, EventId, SerializedGraph } from "../types";
 
-// Re-export GraphEvent for use by other modules
-export type { GraphEvent } from "../types";
-
 /**
  * Event graph for storing operation history
  * This is what gets persisted to disk
