+++ packages/eg-walker/src/graph/topological-walker.ts
@@ -4,7 +4,7 @@
  */
 
 import type { EventId } from "../types";
-import type { GraphEvent } from "../graph/event-graph";
+import type { GraphEvent } from "../types";
 
 export interface EventGraphWalker {
   /**
